Add a file for bash funcitons
Some things cannot really live as separate scripts, as they may want to
change things within the current shell, e.g. the current working directory
or the environment variables.

In this verison, only one command is provided: mkcd, which allows
the user to create a directory and move into it in one action.